refactor(native): split runtime parallel modules and tests
