+++ tests/native/runtime/parallel/contract/runtime_parallel_contract_test.c
@@ -1,5 +1,5 @@
-#include "runtime/parallel/parallel_contract.h"
-#include "runtime/parallel/parallel_model.h"
+#include "runtime/parallel/contract/parallel_contract.h"
+#include "runtime/parallel/model/parallel_model.h"
 
 #include <stdio.h>
